profetisa/requirements.txt
llama-cpp-python
fastapi
uvicorn
python-dotenv
websockets